+++ src/renderer/sdl/audio.rs
@@ -1,5 +1,6 @@
 use super::Renderer;
-use crate::audio::AudioRenderer;
+use crate::audio::{AudioRenderer, AudioStatus};
+use sdl2::audio::AudioStatus as SdlAudioStatus;
 
 impl AudioRenderer for Renderer {
     /// Add audio samples to the audio buffer queue.
@@ -11,5 +12,38 @@ impl AudioRenderer for Renderer {
             std::thread::sleep(std::time::Duration::from_millis(10));
         }
         self.audio_device.queue(samples);
+    }
+
+    /// Return the status of the current audio device.
+    #[inline]
+    fn audio_status(&self) -> AudioStatus {
+        self.audio_device.status().into()
+    }
+
+    /// Return the sample rate of the current audio device.
+    fn audio_sample_rate(&self) -> i32 {
+        self.audio_device.spec().freq
+    }
+
+    /// Resume playback of the current audio device.
+    #[inline]
+    fn resume_audio(&mut self) {
+        self.audio_device.resume();
+    }
+
+    /// Pause playback of the current audio device.
+    #[inline]
+    fn pause_audio(&mut self) {
+        self.audio_device.pause();
+    }
+}
+
+impl From<SdlAudioStatus> for AudioStatus {
+    fn from(status: SdlAudioStatus) -> Self {
+        match status {
+            SdlAudioStatus::Stopped => Self::Stopped,
+            SdlAudioStatus::Playing => Self::Playing,
+            SdlAudioStatus::Paused => Self::Paused,
+        }
     }
 }